Reduce generate exe dependencies
- hashfile no longer run on web files, etag empty
- compute mime type later
- split out html json encoding to own module
- fix bug where all.min.js and .css were not being generated because .gz was removed
