+++ databrary.cabal
@@ -121,15 +121,12 @@ Executable generate
   build-depends:
     aeson >= 0.11.1.4,
     base >= 4.8.2.0,
-    blaze-html >= 0.9.0.1,
     bytestring >= 0.10,
-    cryptonite >= 0.15,
     directory >= 1.2.2.0,
     filepath >= 1.4.0.0,
     hashable >= 1.2.1,
     http-types >= 0.9,
     memory >= 0.8,
-    mime-types >= 0.1.0.6,
     mtl >= 2.2.1,
     parsec >= 3,
     posix-paths >= 0.2.1.0,
